requirements.txt
wandb
tqdm
langchain
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
faiss-cpu
langchainhub
langchain-groq
groq
datasets
ragas
rouge
bert-score
evaluate
tqdm
numpy
pandas
torch
rouge
bert_score
evaluate
rouge_score
python-dotenv
langchain
toml

sentence-transformers
uvicorn
transformers
datasets
